+++ components/containers/Card/BlogCard.js
@@ -1,6 +1,0 @@
-import { CardInterface } from './CardInterface';
-import { BlogCardStyles } from './styles';
-
-export function BlogCard({ card }) {
-  return <CardInterface card={card} styledComponents={BlogCardStyles} />;
-}